improvement: streamline array-mutation ops in ChatForm.tsx
